Indexer: homing and move velocity keep paramInterface busy
When a longer "function" is initiated via the param interface,
the param interface should stay in "busy".
This is valid for homing and move with constant velocity,
See PARAM_IDX_FUN_MOVE_VELOCITY and PARAM_IDX_FUN_REFERENCE.

In this case the param interface does not change the value to DONE
before we are done.