ethercatmcIndexer/Axis.cpp: More fixes for moveVelocity
The param interface is used to trigger an EPICS JOGF,
"move with constant velocity".

Even here we use the param interface. After commanding the function
MOVE_VELOCITY the param interface stays in "BUSY".
When JVEL is changed while jogging, andother moveVelocity is commanded,
while the first one is still active.
There is no (yet) defined way to "change" the velocity.
We need to verify the with the MCU Sw developers.
For the simulator, we stop and command a new motion, making TC 950 pass.

While there, change paramIndexIsFunction() into
paramIndexIsMovingFunction() because "setPosition" is a function,
but does not keep the param interface busy.