Add ethercatmcStartPoller()
Add a iocsh command: ethercatmcStartPoller:
ethercatmcStartPoller("$(MOTOR_PORT)", "$(MOVINGPOLLPERIOD)", "$(IDLEPOLLPERIOD)")

To be useful, the controller must be created with MOVINGPOLLPERIOD and
IDLEPOLLPERIOD set to 0
After the controller is crated, axes can be created, asyn parameters
can be created and the corresponding records can be loaded.

Once all this in in place, the poller can be started.
(Did I mention, that this is mostly useful for PILS, where we get
 an enumberation of devices that want to be exposed as EPICS Records?)

This new function is not used yet, that will be part of the next commits.