Indexer: move velocity does not keep paramInterface busy
Revert commit bcb4ea82:
  Indexer: homing and move velocity keep paramInterface busy

  When a longer "function" is initiated via the param interface,
  the param interface should stay in "busy".
  This is valid for homing and move with constant velocity,
  See PARAM_IDX_FUN_MOVE_VELOCITY and PARAM_IDX_FUN_REFERENCE.
  In this case the param interface does not change the value to DONE
  before we are done.

Having the paramInterface busy while jogging has (at least) one problem:
When the IOC is restarted, it re-connects to the MCU and reads
all parameters, which doesn't work.
In effect, the IOC does not come up.

Talking to the MCU/TwinCAT and PILS experts:
Is is allowed to set the paramInterface into done, when a jog is started.
The motor state (5008/5010) must be busy (which it is).

Start to change the simulator.
Add changes to ethercatmcIndexer to handle this change.
The next commits will be to add test cases.