ci-scripts-build.yml: Use new cache/key after one week
The github-cache is useful to speed up the build:
It keeps a copy of precompiled EPICS-base, asyn, motor (and other modules).

However, the cache is, once created, not changable (on the remote side).
But once the cache is retrieved to local disk, local changes on disk
could be done (like `git fetch origin && git checkout origin/7.0`

But there is no way to push these changes back (and re-use them).

Example:
When the latest from the 7.0 branch from EPICS base is in the cache,
it is there. New commits on this branch are ignored.

Solution, kind of incomplete:
Each cache is identified by a unique "key".
Use a "key" that changes with the number of the week.

Thanks to Ralph Lange for increasing my understanding with 1000%.

And this change is inspired by:
https://github.com/actions/cache