Add ethercatmcDefPos
First step to achive "move to predifend positions"
Add a template that hold the positions, a sel record to
select one and a calc record as an entry for the user.

TODO: The .OMSL and .DOL fields of the motorRecord must be initialized.
An exampe will be test case 800, which will be commited in a second.