ethercatmcDefPos.template: Remove the INPLx fields
It seems as if it is OK to "push" the values ".A .. .L" directly
from the ao records into the sel records.

In that sence: Remove the definition of Iinput links.

Alternative:
Define them to point to the ao records, and remove the OUT field
of the ao records, so that they don't point anywhere.
This seems to be more "sel-Recordish" ?