test/run-ethercatmc-ioc.sh: add --no-make --no-run
Make it possible to run the script in one go or 2:
Today:
  Step 1: run make, generate IOC start script(s)
  Step 2: Execute the IOC start script

Make it possible to run either Step 1, Step 2 or both.

This will be used in the next commit:
In automated testing sometime the make in step 1 takes to long time,
and the test fails.
Now it is possible to run make, and then start all test test cases.
That will be done in the next commit.