test: Add 970_Simulator-MRES_VELO_BVEL_VBAS_VMAX_Sxx_.py
Add pytests36/970_Simulator-MRES_VELO_BVEL_VBAS_VMAX_Sxx_.py
to test if the motorRecord changes VELO or S after changing UREV or MRES.
(in that sense the test could have been named Record)
The "driver uses EGU" feature now demands that VELO is constant
and S is changed, while the traditional handling is that VELO is changed
and S kept as it is.

Note:
For "driver uses EGU" the readback position RRBV does change as well
there is today no test code.