ethercatmcIndexerAxis::addPollNowParam: Don not add twice
When the initial poll was interrupted because
indexerReadAxisParameters() returned in the middle of everything
with an error, the whole procedure was restarted later.
First all other axes where polled, and then our one.

addPollNowParam() was called again with the same parameter index,
and it was happily added to the list again.
Prevent this and check if the index is already in the list.