ethercatmcIndexer.cpp: Improve indexerReadAxisParameters()
Fix a (possible) bug that the method returned early
when reading of a paramer/funtion failed.
This iw wrong.

Print which parameters that are not polled.
Add an early range-check: Make sure that the permission
array does not overflow.
This should never happen, but we never now.
Just make the code robust for future changes.