Add -CfgHVEL-RB
Add the "Homing Velocity read back" to ethercatmcIndexer.h/cpp,
the simulator, to the DB files and the Config Screen.

This is handled similar to JVEL-RB:
Read it from the MCU (if the parameter is there), put it into
the -CfgHVEL-RB PV and forward this into the .HVEL field of the
motorRecord.

Note: What I found in ethercatmcaxisConfig.opi:
Most widgets are TextInput, should probably be TextUdate.