Simulator/indexer: Add usleep
Make it possible to simulate a "slow" TCP connection.
For the moment we want to debug a situation where the connections
times out. This can be done by setting a sleep time > 2 seconds.

The simulator resets the sleep time, when a new connection is accepted.

More test cases will follow.