ethercatmcIndexer.cpp: Handle encoderRaw#2 correctly
The encoderRaw is special:
In the MCU it is typcally an integer value, 16/32 or more bits.
In EPICS, it is a float64, which can hold integer values up
to 48 bits without loosing information.
(and we can run an old asyn without 64 bit integer variables)

The EPICS asyn type for encoderRaw is float64, and in the MCU
we have an int16/int32/int64 type.

Therefor the encoderRaw has special treatment, which did handle
encoderRaw but not the "new form" encoderRaw#2.

Move the parser to the right place.