test/pytests36/980: More tests
Add more tests, since there are more "network problems" in the
simulator.

Add one last test:
Send and invalid network problem to the simulator and check
that it is "rejected".
The simulator returns "Error", and the "-DbgStrToMCU" PV goes
into the "INVALID WRITE" alarm state.

If the simulator ever gets more simulated network problems,
this last test case will fail and TC 890 needs to be updated.