ethercatmcIndexerAxis: Correct INVALID STATE for the UTC variables
The driver was (and is) setting the "INVALID STATE" alarm on the
ai records giving the postion (or raw encoder) timestamped in UTC.
This logic was slightly broken:
When the state machine is e.g. in "START" there is no valid
information, for a short moment.
This did trigger the "INVALIS STATE" alarm.
But it should not. Correct this irritating error.

PS: There may be more work needed for the RESET state,
do that later if needed.