Prepare for 1A04
Prepare for an Discrete Input Device, e.g. 1A04
So far the value can be read.
The status is not handled yet.
Major rework is needed, since we have AUX bits with names,
and they will collide with the AUX bits of an axis.
Unless we treat it like an axis.
More investigation needed.