test/141: More combinations for the NTM test
After a report from external, we found that moving to a new position
(by writing to the .VAL field) and the back to the very start position
(the original .VAL field) did not work for NTM=NO.

Add a test case for this.

Additionally, motorRecord 7.07 has an NTM==UPDATE feature.
Test this as well, when the controller has it.
(See the previous commit, the simulator supports it)