test/920: Fixes for 920: motor master vs ess-master
Make test cases in 920 pass with both motor/master and motor/ess-master
Different fixes/improvements.

One tweek in AxisMR to change the behviour dependent on the motor version.
These kind of tricks should not be there in the ideal world.
However, it is better to have this version-dependency for cases that
are really corner-cases in practice.
Better to have a passing test suite and look into the corner-case later.