test/AxisCom: Introduce putDbgStrToLOG()
Sometime strings > 40 bytes are send to the "-DbgStrToLOG" PV.
Different versions of pvapy (?) seem to handle this differently:
While the strings are silent truncated here on the local system
the latest of all running in the ci pipeline "throws" and error.
The test case fails.

Until there is a realy good handling of longer strings is in place,
the debug log string is put into a try- exept block.