test/122, 132: More tests to run into the limit switch
Major rework of test cases 122 and 132, which run into a limit switch.
The test which "starts jogging" and "once moving disable the soft limits
has been commented out. Setting a parameter while the motor is moving
is not a supported use case in the new MCU SW.

Beside that, add other test cases:
- Use moveVel (outside the motorRecord) instead of JOGF/JOGR via
  the motorRecord
- Set a soft limit to a very high (or low) value instead of disabling
  it.
  This will test the "parameter interface" better:
  Do not rely on what we read back from the MCU, but verify that
  the parameter change works.
  (And more test cases may need to be written)

While there, reformat AxisMr with black