Remove DbgStrToNC
Remove a debug record which allowed to communivate with the NC block.
This kind of debug should not be used any more.
Remove support for Streamdevice: in readOctet() and writeOctet()

The not-used-anymore method readOctet() is removed.
writeOctet() is heavily cleaned up.

If really needed, the code can be retrieved with help of Git later.