ethercatmcIndexerAxis: add pollPowerIsOn(); split poll() into 2 parts
Factor out the core off the poll() method into doThePoll().
Add a parameter, if the data should be read from the process-image
updated by poll() in the controller ("the poller") or if
fresh data should be get from the controller.

Use the "get fresh data" inside pollPowerIsOn(), which makes this method
much shorter