ethercatmcADS: Forward MACHINE/PORT not found error
When the start script is mis-configured, a TCP connection will be
established.
It can happen that rhe ADS route can not be found, because the script
uses the wrong AMSnetid.
Put the error code 0x6 and 0x7 into PV that shows other networks problems
or the "PLCname" once everything is running.