test/run-ethercatmc-ioc.sh: Remove manipulation of ams net id
When the AMS net id is specified in the startup script, do not
manipulate it with help of `sed`.
The same is for the IP-address:
Do not replace 172.x.y.z with 127.0.0.1
This does no longer allow an overwrite on the command line-
when this is really needed, we need to find a better solution.