ethercatmcIndexer.cpp: Improve error logging for param interface
When read-only parameters are written, this is an error in PILSv3.
PILS V2 does not tell us, which parameters are "read" and which are "write",
so change the permissions to "read" once we see PARAM_IF_CMD_ERR_READONLY

In any case, getting a PARAM_IF_CMD_ERR_NO_IDX is a serious error both
in V2 and V3, so keep the "permissions as is" and return an error.