PILSv3: Add lenInPlcParaFloat, lenInPlcParaInteger
Next step towards the more flexible param interface of PILSv3.
Record the length and type of the paramters per axis.
Either it is floating point, then lenInPlcParaFloat is != 0,
or it is an integer parameter (including enum) then
lenInPlcParaInteger is != 0
This information is not used yet, that will be done in the next commit.