PILSv3: Need to read the param descriptor for a function
A "special function" comes with a parameter descriptor: This one
provides information about the parameter type.
Read the descriptor and use the parameter type, if and only if, there
is one.
Example: function "home" may not have a paramter,
but "move with constant velocity" should have.

More information like "min_value" and "max_value" may be handled in
a future commit.