pytests36: Do not test different homing procedures is 110
Phase out the writing of a homing procedure from the IOC:
The homing procedure must be configured in the MCU, and
nowhere else.
It has all definitions, the numeric values as well the the name for it.
In other words remove testing different procedures.
Keep the test to home while sitting on a limit switch