pytests36/141: Improve calculation of timeouts
First of all, add more prints when the motor is stopped because
of a new write to the VAL field and NTM is "active"
Second, reduce the velocity, this may be needed for a motor running
100mm/sec.
Third, increase the timeout when doing NTM tests:
The current timeout may be too short, depending on the velocities of
the motor and the distance.
Calculate a larger timeout: Be prepared that the motor travels the whole
range forth and back.
This is much more then needed, it should be on the safe side.