ethercatmcIndexerAxis: Show nErrorID if != 0
Change the logic to handle cases where the errorID is != 0,
but bError is false.
(If this combination is a good one, is currently under discussion)

Anyway, if we run into this combination, change the "E:" into a "W:"
and display the error as usual.

Move the code to set drvlocal.dirty.old_ErrorId to the right place
And while there, add more setAlarmStatusSeverityWrapper() calls