ethercatmcIndexerAxis: Handle acceleration is read only
When accelation can not be written to the MCU (because it is a
read only parameter) continue with the move() command.
Note the the permission is read only and do not try to
write it again.
Do the same for the velocity in the move() command.
And the same for acceleration in the moveVelocity() command.

Note: The current setting of the (new) permission may be moved
into ethercatmcIndexer.cpp as a cleanup in a later commit.