test/201: ErrorId is set but not Err
Handle the new tc_mca_std_lib better:
Trying to move a motor when power is off (=disabled)
does not set the bError bit (any more).
The nErrorId is set (should be 0x4260)

Handle this case as passed.
Add more debug into the IOC log.
Add a protection when a command "hangs":
when it goes never out of BUSY.
Needs to be investigated on the TwinCAT side.