test/pytests36/AxisMr.py: Add waitForValueChangedInt32
For the next test we need to handle a problem with int32/uint32:
A value like 0x80000000 is an unsigned integer in python,
but a signed one (with a value < 0) in an EPICS longin record.
So that the comparison of the values in waitForValueChanged()
fails.

In short: Write a new method that uses integer and only looks
at 32 bits for comparison.