ethercatmcIndexerAxis.cpp: Update RBV_TSE even if not PTPallGood
One observation while running the long tests:
When the PTP synchonization is "not good", no motor positions had
been updated. This is not ideal, the intention should be the following:
When the motor is not homed, the position is 0.0, STATE INVALID
When the motor is homed, but PTP not synchronized, the position
is updated, the record is put into "STATE INVALID" as well.