Need to call callParamCallbacks(axisNo)
When an additional PILS device was found on a "axisNo" where
no axis was connected, the value in the record had not been updated,
since no callbacks for this "asyn list" had been called.
Fix this by adding a bitmap ctrlLocal.callBackNeeded, set
the corresponding bit and call callParamCallbacks(axisNo) at the
end off poll().