test/simulator: Add fActPosMove
Add a new simulator command to support a flaky 910 test case:
The motor is started "move with constant velocity" and commanded
to a fixed position.
(Techically, it is the other way around)
Once stopped, a backlash command should be issued from the state
machine in the motor record.
The whole set of commands into the axis are recorded into a file
and the file is compared with a pre-computed version.
To make things reproduceable, the motor stop postion is sned into
the simulator.
And here the problem is: The usage fActPosition of will of course
stop the motor.
Depending of the timing of all the polling, the motor record will
not pick up that the move velocity had stopped (after start) and will
not command the backlash.

Solution: Introduce a new command:
"fActPosMove" ("fActPosKeepMoving" would be a better name),
that keeps the motor in the "busy" state (but without moving),
until the stop is commanded (as a consequence of setting
JOGF/JOGR to false.

The needed change in TC910 will be done in the next commit.