ethercatmcIndexerV2.cpp: Set asynDisconnected on aux bit names, if needed
When a PLC reboots, and previous defined AUX bits are no longer defined,
the name was not cleared.
This was observed when moving an aux bit from one position,
and it was still visible.

A better solution would probably to invalidate them already
on disconnect, but that is for later.