test/AxisMr: use repr string in calcAlmostEqual()
Somewhat difficult to explain ?
The printing in python via an f-string will print the "whole object".
For p4p this is the value of a PV and the timestamp (which is correct).

When comparing values of expected and actual, we only want to print the value.
Use the correct "!r" construction to print, like this:
act={actual!r}
Note: There may be more places that could get an "!r"