ethercatmcIndexer.cpp: handle status for 1B08 and friends
When an "input device with status word" is used -
like 1A08, 1B08 or others - the status is used to put an EPICS record
into the alarm state (and severity).
This has only be used when AUX bits are defined, which they often are.
Make it possible to handle the status even without any AUX bit.