Add RAWENCSTEP_ADEL, RAWENCSTEP_MDEL
When a high-resolution encoder is used, it may report a new
value (=encoder tick) very often.
Introduce new macros in ethercatmcIndexerAxis.iocsh to controll
ADEL and MDEL for the encoder raw value.
Change all start scripts, to set this value to 0, except for one
known axis,
More adjustments may be needed with real hardware later

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmc.template
    modified:   iocsh/ethercatmcIndexerAxis.iocsh
    modified:   test/startup/st.ESTIASELENE-pils.iocsh
    modified:   test/startup/st.YMIR-SpRot-MC-Rz-01-MC-MCU-002.iocsh
    modified:   test/startup/st.YMIR-SpScn-MC-MCU-001.iocsh
    modified:   test/startup/st.mcu007-PTP-B02.iocsh
    modified:   test/startup/st.mcu010-indexer.iocsh
    modified:   test/startup/st.mcu015-pils.iocsh
    modified:   test/startup/st.mcu024-pils.iocsh
    modified:   test/startup/st.mcu026-pils.iocsh
    modified:   test/startup/st.mcu028-indexer.iocsh
    modified:   test/startup/st.mcu4004-003-PILS.iocsh
    modified:   test/startup/st.notb-mcs1.iocsh
    modified:   test/startup/st.sim-indexer.iocsh