Prepare handling of InterlockFwd/Bwd
Prepare to handle InterlockFwd and InterlockBwd aux bits better.
Dig out the bit masks for both, without doing anything yet.
Add asynParameters for both
(I wonder if we need them at all after the last changes for auxbitasbirecord ?)
The next step would be to display the Interlock in the MsgTxt