ethercatmcIndexerAxis: Or InterlockFwd/Bwd with limit switches
Not fully ideal, but probably a step into the right direction:
Or the physical limit switches with the Interlock bits, and
feed those into the motor module.
This will e.g. prevent the motor from trying to do retries,
or powering on a motor which is interlocked.
The problem for the momemt is that the debugging is harder,
especially in the UI.