StatusBits-alarms.txt and code
Make it easier to use the reason bits and status word defined in PILS
to set the alarm status on EPICS records.
The status word RESET, DISABLED, WARNING, ERROR together with the reason
bits in a "discrete input device (0x1a04) or "analog input" (0x1b04/0x1b08)
are ment to have a 1:1 relation to the .SEVR and .STAT fields of the
corresponding "longin" or "ai" record in EPICS.

The different systems have slightly different definitions, an 1:1
mapping is not really possible.
For now, treat ERROR always as hardware ERROR.

Remove HIHI and LOLO. (if needed we can have an AUX bit for that).

Support the DISABLED state: It means that the sensor is "masked" inside
the MCU. The value is no longer connected to the sensor, and may be patched
online to simulate and test the behaviour when e.g. a temperature is changed,
gets too high and triggers a "motor disabled" state.
Or just test the alarm handling in the different control systems,
MCU, EPICS, CSS, NICOS.
Changes to be committed:
    modified:   docs/AlarmState-MsgTxt.txt
    new file:   docs/StatusBits-alarms.txt
    modified:   ethercatmcApp/src/ethercatmcController.cpp
    modified:   test/pytests36/940_Simulator-InputRecordAlarm.py