ethercatmcIndexerV2: Fix reading of info type 0
commit 1bf7825a1c1e5e568e0,
"ethercatmcIndexerV2: Limit the read length of e.g. names", introduced a
problem that had not been noted until today:
The length of the indexer is unknown, and reading it did not work,
since the read len was clipped to 0.
Solution: limit the length to be read only when the indexer size is known.