st.mcu035-PTP-B02.iocsh Add a sleep to let the poller read the aux bit names
Add a workaround for the race condition when reading the aux bit names
the same time that the records are loaded caused by iocInit.

In prociple, this is not a race-condition, since there is a proper
handling: Define(=create) the needed parameters in the iocsh file.

This will be done in the next commit.