ethercatmcEL6688-PTP.template: Remove the "rewrite offset"
Remove 2 records that dealt with the offset of the EL6688.
This value (and the function to set it to 0) is no longer used.
Inside the MCU we calculate the "synched" state with help of a FB,
that checks for a (configurable) tolerance of 5000 nsec.
(and that is what we should have done from the beginning)