test/run-ethercatmc-ioc.sh: Add support for --epics-twincat-ads
Add support to use ethercatmc, the non-indexer version) together with
m-epics-twincat-ads as a transport layer:
The first version of ethercatmc (this module) uses a TCP socket to
communicate with the motion controller.
And it assumes that the motion controller has the right firware version,
tc_mca_std_lib in the first generation:
The one that uses MAIN.M2 for motor 1
Later epics-twincat-ads had been developped, which uses ADS instead of an
ASCII command set via the TCP connection.
Add the option
--epics-twincat-ads
to let ethercatmc talk to MCUs running the MAIN.M1 version use ADS