Add ethercatmcEL6688-PTP-twincat-ads.template
Make it possible to run an IOC with EL6688 using the epics-twincat-ads
communication between the IOC and the "motion controller".
Note that the current SW combination
epics-twincat-ads and tc_mca-std_lib has no support for motion.
The name "motion controller" is in that sense somwhat miss-leading.

Anyway, copy and convert ethercatmcEL6688-PTP into
ethercatmcEL6688-PTP-twincat-ads.template to have allow that the
"mini-PTP-crate" can be uses to test epics-twincat-ads instead of
ethercatmc with PILS.