ci/checkshellcheck: Double quote ... SC2086
Address the warning
"Double quote to prevent globbing and word splitting. [SC2086]"
When shellcheck was introduced, it was somewhat missed that there
where a ton of warnings.
Start to address the most common:
"Double quote"

There are 3 ways to adsress this:
- (1) Rewrite the code
- (2) Disable the warning on a file-by-file bases
- (3) Disable the warning on command line

This patch goes for (1) when there is only one or two lines to be changed,
and (2) when there are many lines to be changed.
(2) is choosen to avoid touching working shell-code with the risk that
something brakes: Not all shell scripts are run on a regular bases.
It is better to look at these when the file is changed the next time.

And as a side-note: More wanings need to be addressed in the next commits