PTP: PTPdiffNTtime_MCU, part 1
Implement a new functionality:
Calculate and show the difference between the PTP and the NTP
time in the motion controller unit ("MCU").

What is tis good for ?
It can be used to check and verify that the different components of the
overall timing system are working.

EVG: (ESS) Event Generator. The master of facility time and all events
           driving the accelerator and other equipment.
           EVRs are connected through a time-compensated optical network.
EVR        Event receiver. Hardware that receives the modulated
           88Mhz signal from the EVG. Can extract the time.
           Can generate 3.3v (or 5V) pulses for external equipment.
           Can timestamp pulses from external equipment.
           Can be equipped with a PTP master, which an distribute
           the "facility time" via a PTP network.
MCU: motion control unit.

EVG  -> timeing network (fiber) -> EVR -> PTP master -> ptp slave in MCU
EVG  -> NTP master    -> network ->                     ntp slave in MCU

This commit handles the NTP time from the MCU.
Technically it is only available as "broken down time":
year, month... seconds, milliseconds.

As there is now function inside the MCU to convert it into epoch seconds
(and the nanoseconds), this information is fiddled into a 64 bit integer
and de-fiddled in the IOC (this code).
gmtime() is used to convert into epoch - and then we convert into EPICS epoch.

One change that is part of the re-construction:
Use ctrlLocal.systemUTCtimeFunction
  The existing variable ctrlLocal.systemUTCtimeOffset was only used
  to indicate that there is a variable "SystemUTCtime" inside the MCU.
  Which allows to timestamp the motor positions in MCU time.

  However, in the next commit the function is needed, to compare the UTC
  time in the MCU (which is assumed to be PTP) with the NTP time of the MCU:
  Introduce a systemUTCtimePTPFunction variable

The new code uses the already existing function epicsTimeDiffInSeconds(),
something that can be used to calculate the difference between PTP in the MCU
and NTP inside the IOC ("PTPdiffTimeIOC_MCU_").
More cleanups (and one more time inside the MCU) will be added later.