Add .clang-format
Add .clang-format to make it possible for auto-formating of all
c/c++ code.
Stick to one of the defaults that clang-format offers.
Seems as if the Google version is slightly closer to the EPICS/asyn style,
so go for this one instead of llvm

clang-format-mp-14 -style=Google -dump-config > .clang-format

No code changes yet.