test/simulator: Support fSimForcePos
Add a new function inside the simulator to move a motor very fast
to a desired position:
fSimForcePos

The difference to fActPosition is that fSimForcePos just moves the
motor, in opposite to fActPosition, which sets "a logical" position,
moving the hard- and soft limits in hardware.
Example:  the driving range is -19.0 .. +18.0
Using fSimForcePos(18.0) will activate the high limit switch.
Using fActPosition(18.0) will not:
The new driving range is now -1.0 .. +36.0

In other words, it is a fastmove.
Improve the way the "activeted limit switch" is calculated in
getPosLimitSwitch() and getNegLimitSwitch(): Do not wait one cycle,
but calculate the actual position