test: Add 921...HLS_LLS_do-not_move-further
Add a new test case to see what is happening, when the motor
sits on a limit switch, and the user tries to "push it further".
The original motorRecord does seem to try do do so,
and does not always suppress the move, but sends it to the controller.
Investigations ongoing