PTP: Add PTPdiffTcNTPExttime_MCU
Add handling for yet-another-PTP-NTP diff:
The TcNTPExttime against PTP.
This will compare the time from an NTP client inside the TC runtime
to the outside world (= an NTP server of course)
In opposite to existing diffs, where either
a) The (NTP-synched) time of the IOC is used:
   This is slighty wrong because the polling between the IOC and the MCU
   It may even be "wronger", when the IOC is run in a virtual machine,
   where the NTP client is inside the virtual machine
   (and not the host machine)
b) The (NTP-synched) time inside the MCU host system:
   Either Windows or FreeBSD.
   There will be, similar to a) always a delay of some milliseconds,
   caused by the PLC cycle time and using ADS to talk to the outside
   world (of TwinCAT)
c) The TC time, inside TwinCAT itself. Initialized at start,
   and the running free.
   There is a difference between the NTP clients in Windos and BSD.
   And running free is not the case, when we have a proper configured
   PTP terminal, which is configure to sync Distributed Clock, so that
   it does not (slowly) drift away from the outside world.

If this feels complicated: it is. Search for TwinCAT time sources:
https://infosys.beckhoff.com/english.php?content=../content/1033/ethercatsystem/2469114379.html&id=

Did some more changes in op/Boy, to generate the new OPI files.
And of course, you need the new firmware inside the MCU.

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmcEL6688-PTP.template
    modified:   ethercatmcApp/src/ethercatmcController.h
    modified:   ethercatmcApp/src/ethercatmcIndexer.cpp
    modified:   ethercatmcExApp/op/Boy/ethercatmcPTPErrBits.opi
    modified:   ethercatmcExApp/op/Boy/tools/Makefile
    modified:   ethercatmcExApp/op/Boy/tools/genExpertWithAuxBits.sh