Change MOTOR_NAME into AXIS_NAME
Cleanup the macros:
MOTOR_NAME is replaced by AXIS_NAME, which will make it consistant for
"pneumatic axes".
While it is debatable if a pneumatic axis is a motor or not,
the same infrastructure (asyn parameters) is already used today.
See handling for a 5010 device and 1e04, they are both axes,
with a cmd/status word, aux bits and everything.