readEnumsAndValueAndCallbackIntoMbbi()
Re-do the reading of "enums" for mbbi records:
The callback has to be postponed until the global variable
interruptAccept in EPICS base is not 0 (any more) and all records
are ready to accept interrupts.
Thanks to the experts on techtalk for helping here.

Re-order all the realted code into a new method:
readEnumsAndValueAndCallbackIntoMbbi()
This change will make the enums for a 1e04 device work again.