Replace PREFIX with SYSPFX, WIP
Start to seperate the PREFIX, to allow using the ESS naming convention.
Some Records have just one macro, $(PREFIX), and are used for the MCU,
like e.g. the PTP stuff and later temperature in the rack or so.
They will go into a diffrent name space.
Axes will have a different name space as well.
More work to be done.