test/run-ethercatmc-ioc.sh: move always old logfile to log/
The run-ethercatmc-ioc.sh script can be run with the -l option
to tee messages (printouts from the IOC)  to a file.
However, when run without -l, the logfile is left there, leading to
confusion if someone thinks that this is the latest one -
but it is not.

Solution:
Move the old logfile always into log/