op/Boy: More for shutters at ymir-mcs1
Do a lot of changes to display 3 shutters at YMIR.
Re-do the macros (use R1 instead of R9) so that
more than one shutter can be displayed, similar to M1

Increase the shutter number in tools/genMotorShutter.sh
See tools/Makefile, how ../motor-ymir-mcs1-temp.opi is created

There may be more things to be cleaned up later,
like $(P) vs $(PREFIX) and other macros used in the iosh files vs
the opi files. That needs to be done in a different commit.

Changes to be committed:
    modified:   IOC.opi
    new file:   motor-ymir-mcs1-temp.opi
    modified:   shutter-1.opi
    modified:   tools/IOC-MCU032.mid
    modified:   tools/IOC-Shutter.mid
    modified:   tools/IOC-YMIR-MCS1.mid
    modified:   tools/IOC.start
    modified:   tools/Makefile
    modified:   tools/genMotorShutter.sh
    modified:   tools/shutterx.mid