ethercatmcAxis: No error when axis is not homed
A not-homed motor should signal this to the motorRecord
in some way, to trigger an alarm state.
The upstream version of motor needs to set the error bit,
(even called PROBLEM).
The ess-master version has the motorNotHomedProblem feature.
Use it, to avoid a moving motor with problems, making motorRecord
think that this is an "external move"