op/Boy: Split IOC-MCU010 into -ads and -pils
Make it possible to use the crate 010 for both -ads and -pils:
This is tc_mca_std_lib in the MCU:
version v1.0.0-6 using the epics-twincat-ads driver for EPICS
versus the v2.x.y versions that use the PILS protocol.
Note: We not expect any big developmwnt here, this is just to demo
and teaching.