test/simulator: Prepare logging with time stamps
Try to debug, why/when setMotorPos() was called in the simulator.
As a first step, add a new file logerr_info.c which has time-getting
printing code.
This was copied from a similar function from EPICS base (the IPv6 project)

Add a new macro, LOGTIME3().

It will be used in the next commit.
Changes to be committed:
    modified:   test/simulator/Makefile
    new file:   test/simulator/logerr_info.c
    modified:   test/simulator/logerr_info.h