test/921: Always initialize
Break out the initializion, so far done once, and run it, if needed,
in each and every test case.
This will make debugging much easier in the coming changes,
where test case will be run one-by-one.
Re-number the test cases, introduce a '0', to give room for more
test cases later.
All in all: no intended code changes.
TODO: fix MRES, it is not written as intended.