test: Various imporvements for 921
Start to test the MRES < 0.0 feature.
One (or more) problem needs to be resolved:
When MRES is change inside the motorRecord, it will send us a "load pos",
messing up the hardware configuration of the simulator.
This "load pos" is only done, when the drive does not use EGU,
but this exactly, what the test wants to test.
Before re-arranging the start/initialization, commit all improvments.

Changes to be committed:
    modified:   pytests36/921_Simulator_HLS_LLS_do-not_move-further.py
    modified:   pytests36/AxisMr.py
    modified:   simulator/hw_motor.c
    modified:   simulator/hw_motor.h
    modified:   simulator/indexer.c