ethercatmcIndexer: Support of named custom parameters
Support reading the name of a custom parameter, if supplied by the MCU.
Side note: This is not an official feature of the PILS v2 specification.

However, it is useful to hava a flexible way to add more parameters
to be read from or written to, by giving them a name.
That name is then used for the asynParameter.

This commit includes the following changes:
ethercatmcIndexer.cpp: Correct the range of old custom paramters
  In the beginning some of the custom PILS parameters had been
    defined as integers, with the motivation that they are boolean.
  However, later all custom parameters where changed into floating point,
  which makes the code easier both in the IOC and the MCU.
  To avoid confusion, the old integer parameters got a new numbering.
  Document the old range, now legacy.

Read and print the name of a custom parameter

make plcParamIndexTxtFromParamIndex() a class function
  In the past, plcParamIndexTxtFromParamIndex() used a hard-coded
  translation of paramIndex into a text representation.
  With the upcoming dynamic handling of custom parameters,
  where the MCU provides a name, this function is now a class method.